backend/requirements.txt
fastapi
uvicorn
authlib
httpx
python-dotenv
itsdangerous
python-jose
